Bunch of refactoring. Just need guest player functionality then we are good enough for prod

+++ internal/club/types.go
@@ -30,5 +30,5 @@ type PlayerInfo struct {
 	ID               string
 	Name             string
 	BallBringerCount int
-	Level            float32
+	Level            float64
 }
